Move market to it's own module - Also create  separate test file - And remove unnecessary exports from tests
