+++ v3/subgraphs/goerli/src/core.ts
@@ -7,12 +7,6 @@ import {
   PoolNominationRevoked,
   PoolOwnerNominated,
 } from '../generated/PoolModule/PoolModule';
-import {
-  MarketRegistered,
-  MarketManagerModule,
-  MarketUsdDeposited,
-  MarketUsdWithdrawn,
-} from '../generated/MarketManagerModule/MarketManagerModule';
 import { Deposited, Withdrawn } from '../generated/CollateralModule/CollateralModule';
 import { CollateralConfigured } from '../generated/CollateralConfigurationModule/CollateralConfigurationModule';
 import {
@@ -26,7 +20,7 @@ import {
   RewardsDistributorRegistered,
 } from '../generated/RewardsManagerModule/RewardsManagerModule';
 import { Liquidation, VaultLiquidation } from '../generated/LiquidationModule/LiquidationModule';
-import { DelegationUpdated, VaultModule } from '../generated/VaultModule/VaultModule';
+import { DelegationUpdated } from '../generated/VaultModule/VaultModule';
 import { UsdMinted, UsdBurned } from '../generated/IssueUSDModule/IssueUSDModule';
 import {
   Pool,
@@ -44,7 +38,7 @@ import {
   Liquidation as LiquidationEntity,
   VaultLiquidation as VaultLiquidationEntity,
 } from '../generated/schema';
-import { BigDecimal, BigInt, Bytes, store, log } from '@graphprotocol/graph-ts';
+import { BigDecimal, BigInt, Bytes, store } from '@graphprotocol/graph-ts';
 
 ////////////////////
 // Event handlers //
@@ -114,25 +108,6 @@ export function handleNewPoolOwner(event: PoolOwnershipAccepted): void {
     pool.save();
   }
 }
-
-/////////////
-// Market //
-////////////
-
-export function handleMarketCreated(event: MarketRegistered): void {
-  const newMarket = new Market(event.params.marketId.toString());
-  newMarket.address = event.params.market;
-  newMarket.created_at = event.block.timestamp;
-  newMarket.created_at_block = event.block.number;
-  newMarket.updated_at = event.block.timestamp;
-  newMarket.updated_at_block = event.block.number;
-  newMarket.usd_deposited = BigDecimal.fromString('0');
-  newMarket.usd_withdrawn = BigDecimal.fromString('0');
-  newMarket.net_issuance = BigDecimal.fromString('0');
-  newMarket.reported_debt = BigDecimal.fromString('0');
-  newMarket.save();
-}
-
 export function handlePoolConfigurationSet(event: PoolConfigurationSet): void {
   const pool = Pool.load(event.params.poolId.toString());
   // Pool will be never undefined, though for safety reasons we are checking for that
@@ -207,53 +182,6 @@ export function handlePoolConfigurationSet(event: PoolConfigurationSet): void {
     pool.updated_at_block = event.block.number;
     pool.save();
   }
-}
-
-export function handleMarketUsdDeposited(event: MarketUsdDeposited): void {
-  const market = Market.load(event.params.marketId.toString());
-  if (market == null) {
-    log.error(
-      'Something went wrong, got a MarketUsdDeposited event for a market that doesnt exists ' +
-        event.params.marketId.toString(),
-      []
-    );
-    return;
-  }
-  const contract = MarketManagerModule.bind(event.address);
-  const reported_debt = contract.getMarketReportedDebt(event.params.marketId).toBigDecimal();
-  const usd_deposited = market.usd_deposited.plus(event.params.amount.toBigDecimal());
-  const net_issuance = market.net_issuance.minus(event.params.amount.toBigDecimal());
-
-  market.reported_debt = reported_debt;
-  market.updated_at = event.block.timestamp;
-  market.updated_at_block = event.block.number;
-  market.net_issuance = net_issuance;
-  market.usd_deposited = usd_deposited;
-  market.save();
-}
-
-export function handleMarketUsdWithdrawn(event: MarketUsdWithdrawn): void {
-  const market = Market.load(event.params.marketId.toString());
-  if (market == null) {
-    log.error(
-      'Something went wrong, got a MarketUsdWithdrawn event for a market that doesnt exists ' +
-        event.params.marketId.toString(),
-      []
-    );
-    return;
-  }
-
-  const contract = MarketManagerModule.bind(event.address);
-  const reported_debt = contract.getMarketReportedDebt(event.params.marketId).toBigDecimal();
-  const usd_withdrawn = market.usd_withdrawn.plus(event.params.amount.toBigDecimal());
-  const net_issuance = market.net_issuance.plus(event.params.amount.toBigDecimal());
-
-  market.reported_debt = reported_debt;
-  market.updated_at = event.block.timestamp;
-  market.updated_at_block = event.block.number;
-  market.net_issuance = net_issuance;
-  market.usd_withdrawn = usd_withdrawn;
-  market.save();
 }
 
 //////////////
